feat(many): update for 1.30.0-2c0be4f
