+++ pkg/ams/shared/util_windows.go
@@ -1,0 +1,31 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+/*
+ * This file is part of AMS SDK
+ * Copyright 2021 Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the Lesser GNU General Public License version 3, as published
+ * by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranties of MERCHANTABILITY, SATISFACTORY
+ * QUALITY, or FITNESS FOR A PARTICULAR PURPOSE.  See the Lesser GNU General Public
+ * License for more details.
+ *
+ * You should have received a copy of the Lesser GNU General Public License along
+ * with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+//go:build windows
+// +build windows
+
+package shared
+
+import (
+	"os"
+)
+
+// GetOwnerMode retrieves the file mode and owner of a given file object
+func GetOwnerMode(fInfo os.FileInfo) (os.FileMode, int, int) {
+	mode := fInfo.Mode()
+	return mode, 0, 0
+}
